Move views model
