+++ ArchipelagoParser/ViewModel/CustomUserControl/RandoSlotViewModel.cs
